+++ config/wdio/wdio.conf.js
@@ -42,6 +42,9 @@ const seleniumGridUrl = process.env.SELENIUM_GRID_URL;
  */
 const browsers = process.env.BROWSERS;
 
+/* Use to override default theme for theme visual regression tests. */
+const theme = process.env.THEME;
+
 const hasPackages = glob.sync((path.join(process.cwd(), 'packages'))).length > 0;
 
 const seleniumConfig = determineSeleniumConfig({
@@ -99,6 +102,8 @@ const config = {
     timeout: 1200000,
     bail,
   },
+
+  ...theme && { theme },
 };
 
 // This code only executes for monorepos.  It will create a set of suites that can then be executed