+++ scripts/wdio/wdio-runner-cli.js
@@ -15,6 +15,7 @@ commander
   .option('--locales [list]', 'The list of locales to test. Defaults to [en]', parseCLIList, undefined)
   .option('--formFactors [list]', 'The list of viewport sizes to test.', parseCLIList, undefined)
   .option('--browsers [list]', 'The list of browsers to test. Defaults to [chrome].', parseCLIList, undefined)
+  .option('--themes [list]', 'List of themes to override defined default theme.', undefined)
   .option('--gridUrl [url]', 'The selenium grid url to run tests against', undefined)
   .option('--continueOnFail', 'Pass to continue executing test runs when a run fails', false)
   .option('--updateReference', 'Pass to remove all reference screenshots during screenshot cleanup', false)
@@ -30,6 +31,7 @@ const {
   config,
   gridUrl,
   browsers,
+  theme,
   formFactors,
   locales,
   host,
@@ -54,6 +56,7 @@ runner({
   locales,
   gridUrl,
   browsers,
+  theme,
   // honored wdio cli options
   ...host && { host },
   ...port && { port },