+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changelog
 ----------
 ### Added
 * Added a post-install console warning if terra-toolkit is included as a hard dependency
+* Fallback theme generation strategy for theme aggregation.
+* Theme visual regression support.
 * Added a line to webpack config to only resolve to 'main'
 
 5.6.0 - (August 30, 2019)